+++ RSS-React/src/App.tsx
@@ -1,4 +1,4 @@
-import './App.css';
+import './styles/App.css';
 import { Route, Routes } from 'react-router-dom';
 import Home from './pages/home';
 import NotFound from './pages/notFound';
