+++ RSS-React/src/components/loader.tsx
@@ -1,4 +1,4 @@
-import './loader.css';
+import './../styles/loader.css';
 
 function Loader() {
   return (
