+++ RSS-React/src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
-import './index.css';
+import './styles/index.css';
 import ErrorBoundary from './services/ErrorBoundary.tsx';
 import { BrowserRouter } from 'react-router-dom';
